requirements.txt
inquirerpy==0.3.4
requests==2.28.1
google-api-python-client==2.89.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
Flask==2.3.2
moviepy==1.0.3
pytube==15.0.0